requirements.txt
pypdf>=3.17.0
python-pptx>=0.6.22
colorama>=0.4.6
